Abstract
Nowadays, information hiding is becoming a helpful technique and fetch more attention due fast growth of using internet, it is applied for sending secret information by using different techniques. Stenography is one of major important technique in information hiding. Stenography is science of concealing the secure information within a carrier object to provide the secure communication though the internet, so that no one can recognize and detect it's except the sender & receiver. In stenography, many various carrier formats can be used such as an image, video, protocol, audio. The digital image is most popular used as a carrier file due its frequency on internet. There are many techniques variable for image stenography, each has own strong and weak points. In this study, we conducted a review of image stenography in spatial domain to explore the term image stenography by reviewing, collecting, synthesizing and analyze the challenges of different studies which related to this area published from 2014 to 2017. The aims of this review is provides  an overview of  image stenography and comparison between approved studies are discussed according to the pixel selection , payload capacity and  embedding algorithm to  open  important research issues in the future works and obtain a robust method.